Remove bad error check from RunMove
The RunMove function checks if a move was accepted that resulted in a
huge increase in energy. This was intended to be a check for a huge
_decrease_ in energy so we could recalculate the energy to avoid round
off. However, this seems unnecessary. And, the current check  may allow
the system to work around any possible bug in the code that would accept
bad moves.