Make the Maximum Displacement a Percent of the Box Axis Length
Change the maximum displacement to be a percent of the box axis length
instead of a fixed value based on the box axis length, with a limit of
just under 0% of the box axis length. This guarantees that the
displacement value selected, after periodic boundaries, will stay within
the box, even for unstable systems.